+++ mediation/LiftoffMonetize/source/plugin/Assets/GoogleMobileAds/Mediation/LiftoffMonetize/Plugins/iOS/LiftoffMonetizeExtrasBuilder.h
@@ -13,11 +13,10 @@
 // limitations under the License.
 
 #import <Foundation/Foundation.h>
-
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
 #import "GADUAdNetworkExtras.h"
 
-@interface VungleExtrasBuilder : NSObject <GADUAdNetworkExtras>
+@interface LiftoffMonetizeExtrasBuilder : NSObject <GADUAdNetworkExtras>
 
 @end